# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BoardItemCategory = typeof BoardItemCategory[keyof typeof BoardItemCategory];


export const BoardItemCategory = {
  idea: 'idea',
  question: 'question',
  update: 'update',
} as const;

export interface BoardItem {
  id: number;
  title: string;
  body: string;
  category: BoardItemCategory;
  authorName: string;
  reactions: number;
  reactedByCurrentUser: boolean;
  createdAt: string;
  updatedAt: string;
}

export type BoardItemInputCategory = typeof BoardItemInputCategory[keyof typeof BoardItemInputCategory];


export const BoardItemInputCategory = {
  idea: 'idea',
  question: 'question',
  update: 'update',
} as const;

export interface BoardItemInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  body: string;
  category: BoardItemInputCategory;
  /**
     * @minLength 1
     * @maxLength 40
     */
  authorName: string;
}

export interface ReactionInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  authorName: string;
}

export type BoardActivityAction = typeof BoardActivityAction[keyof typeof BoardActivityAction];


export const BoardActivityAction = {
  created: 'created',
  reacted: 'reacted',
  unreacted: 'unreacted',
  deleted: 'deleted',
} as const;

export interface BoardActivity {
  id: number;
  itemId: number;
  actorName: string;
  action: BoardActivityAction;
  itemTitle: string;
  createdAt: string;
}

export interface BoardSummary {
  totalItems: number;
  ideas: number;
  questions: number;
  updates: number;
  reactions: number;
}

export interface ErrorResponse {
  error: string;
}