# lib/api-zod/src/generated/types/boardActivity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BoardActivityAction } from './boardActivityAction';

export interface BoardActivity {
  id: number;
  itemId: number;
  actorName: string;
  action: BoardActivityAction;
  itemTitle: string;
  createdAt: Date;
}