# lib/api-zod/src/generated/types/boardItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BoardItemCategory } from './boardItemCategory';

export interface BoardItem {
  id: number;
  title: string;
  body: string;
  category: BoardItemCategory;
  authorName: string;
  reactions: number;
  reactedByCurrentUser: boolean;
  createdAt: Date;
  updatedAt: Date;
}