# lib/api-zod/src/generated/types/boardItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BoardItemInputCategory } from './boardItemInputCategory';

export interface BoardItemInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  body: string;
  category: BoardItemInputCategory;
  /**
     * @minLength 1
     * @maxLength 40
     */
  authorName: string;
}