# lib/api-zod/src/generated/types/boardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BoardSummary {
  totalItems: number;
  ideas: number;
  questions: number;
  updates: number;
  reactions: number;
}