# lib/api-zod/src/generated/types/reactionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReactionInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  authorName: string;
}